Add `no-use-before-define` to `sdk/.eslintrc.js` rules
