+++ src/typescript/sdk/src/emojicoin_dot_fun/index.ts
@@ -2,5 +2,6 @@ export * from "./calculate-swap-price";
 export * from "./contract-apis";
 export * from "./events";
 export * from "./serialize-entry-args-to-json";
+export * from "./post-bcs-view-function";
 export * from "./types";
 export * from "./utils";
