+++ src/typescript/sdk/src/emojicoin_dot_fun/payload-builders.ts
@@ -13,23 +13,21 @@ import {
   type UserTransactionResponse,
   type WaitForTransactionOptions,
   Serializable,
-  Serializer,
+  type Serializer,
   type AnyRawTransaction,
   AccountAuthenticator,
   type InputViewFunctionData,
   TransactionPayloadMultiSig,
   MultiSigTransactionPayload,
   type MoveValue,
-  MimeType,
-  postAptosFullNode,
   type AptosConfig,
   type InputGenerateTransactionOptions,
   type AccountAddressInput,
   type InputEntryFunctionData,
 } from "@aptos-labs/ts-sdk";
-import { toConfig } from "../utils/aptos-utils";
 import serializeEntryArgsToJsonArray from "./serialize-entry-args-to-json";
 import { type StructTagString } from "../utils/type-tags";
+import { postBCSViewFunction } from "./post-bcs-view-function";
 
 export class EntryFunctionTransactionBuilder {
   public readonly payloadBuilder: EntryFunctionPayloadBuilder;
@@ -230,26 +228,4 @@ export abstract class ViewFunctionPayloadBuilder<T extends Array<MoveValue>> {
   argsToArray(): Array<EntryFunctionArgumentTypes> {
     return Object.keys(this.args).map((field) => this.args[field as keyof typeof this.args]);
   }
-}
-
-/* eslint-disable-next-line import/no-unused-modules */
-export async function postBCSViewFunction<T extends Array<MoveValue>>(args: {
-  aptosConfig: Aptos | AptosConfig;
-  payload: EntryFunction;
-  options?: LedgerVersionArg;
-}): Promise<T> {
-  const { payload, options } = args;
-  const aptosConfig = toConfig(args.aptosConfig);
-  const serializer = new Serializer();
-  payload.serialize(serializer);
-  const bytes = serializer.toUint8Array();
-  const { data } = await postAptosFullNode<Uint8Array, MoveValue[]>({
-    aptosConfig,
-    path: "view",
-    originMethod: "view",
-    contentType: MimeType.BCS_VIEW_FUNCTION,
-    params: { ledger_version: options?.ledgerVersion },
-    body: bytes,
-  });
-  return data as T;
 }
