+++ src/typescript/sdk/tests/e2e/queries/market-state.test.ts
@@ -1,5 +1,5 @@
 import { type SymbolEmoji } from "../../../src";
-import TestHelpers from "../../utils/helpers";
+import { registerMarketHelper } from "../../utils/helpers";
 import { getFundedAccount } from "../../utils/test-accounts";
 import { waitForEmojicoinIndexer } from "../../../src/indexer-v2/queries/utils";
 import { SwapWithRewards } from "@/contract-apis/emojicoin-dot-fun";
@@ -17,11 +17,10 @@ describe("queries a market by market state", () => {
 
   it("fetches the market state for a market based on an emoji symbols array", async () => {
     const emojis: SymbolEmoji[] = ["🧐", "🧐"];
-    const { registerResponse, marketAddress, emojicoin, emojicoinLP } =
-      await TestHelpers.registerMarketFromEmojis({
-        registrant,
-        emojis,
-      });
+    const { registerResponse, marketAddress, emojicoin, emojicoinLP } = await registerMarketHelper({
+      registrant,
+      emojis,
+    });
     const { version } = registerResponse;
     await waitForEmojicoinIndexer(version);
     const res = (await fetchMarketState({
