+++ src/typescript/sdk/tests/e2e/queries/num-markets.test.ts
@@ -2,7 +2,7 @@ import { getRegistryResourceFromWriteSet, toRegistryView } from "../../../src";
 import { type SymbolEmojiName } from "../../../src/emoji_data/types";
 import { RegistryView } from "@/contract-apis/emojicoin-dot-fun";
 import { getAptosClient } from "../../../src/utils/aptos-client";
-import TestHelpers from "../../utils/helpers";
+import { registerMarketHelper } from "../../utils/helpers";
 import { getFundedAccounts } from "../../utils/test-accounts";
 import {
   fetchLargestMarketID,
@@ -30,7 +30,7 @@ describe("fetches the number of registered markets based on the latest processed
 
   beforeAll(async () => {
     const registerAll = marketEmojiNames.map((emojiNames, i) =>
-      TestHelpers.registerMarketFromNames({
+      registerMarketHelper({
         registrant: registrants[i],
         emojiNames,
       }).then(({ registerResponse }) => {
