+++ src/typescript/sdk/tests/e2e/queries/search-emojis.test.ts
@@ -1,5 +1,5 @@
 import { type SymbolEmojiName, namesToEmojis } from "../../../src";
-import TestHelpers from "../../utils/helpers";
+import { registerMarketHelper } from "../../utils";
 import { getFundedAccounts } from "../../utils/test-accounts";
 import { fetchMarkets } from "../../../src/indexer-v2/queries/app/home";
 import { waitForEmojicoinIndexer } from "../../../src/indexer-v2/queries/utils";
@@ -25,7 +25,7 @@ describe("queries markets by the various emojis in their symbols", () => {
 
   beforeAll(async () => {
     const registerAll = marketEmojiNames.map((emojiNames, i) =>
-      TestHelpers.registerMarketFromNames({
+      registerMarketHelper({
         registrant: registrants[i],
         emojiNames,
       }).then(({ registerResponse }) => registerResponse)
