+++ src/typescript/sdk/tests/e2e/queries/simple.test.ts
@@ -1,5 +1,5 @@
 import { type SymbolEmojiName, getEvents, ONE_APT } from "../../../src";
-import TestHelpers, { EXACT_TRANSITION_INPUT_AMOUNT } from "../../utils/helpers";
+import { EXACT_TRANSITION_INPUT_AMOUNT, registerMarketHelper } from "../../utils/helpers";
 import { Chat, ProvideLiquidity, Swap } from "@/contract-apis/emojicoin-dot-fun";
 import {
   fetchChatEvents,
@@ -38,7 +38,7 @@ describe("queries swap_events and returns accurate swap row data", () => {
   });
 
   it("performs a simple registerMarket fetch accurately", async () => {
-    const { registerResponse: response } = await TestHelpers.registerMarketFromNames({
+    const { registerResponse: response } = await registerMarketHelper({
       registrant,
       emojiNames: marketEmojiNames[0],
     });
@@ -54,7 +54,7 @@ describe("queries swap_events and returns accurate swap row data", () => {
   });
 
   it("performs a simple swap fetch accurately", async () => {
-    const { marketAddress, emojicoin, emojicoinLP } = await TestHelpers.registerMarketFromNames({
+    const { marketAddress, emojicoin, emojicoinLP } = await registerMarketHelper({
       registrant: swapper,
       emojiNames: marketEmojiNames[1],
     });
@@ -80,11 +80,10 @@ describe("queries swap_events and returns accurate swap row data", () => {
   });
 
   it("performs a simple chat fetch accurately", async () => {
-    const { marketAddress, emojicoin, emojicoinLP, emojis } =
-      await TestHelpers.registerMarketFromNames({
-        registrant: user,
-        emojiNames: marketEmojiNames[2],
-      });
+    const { marketAddress, emojicoin, emojicoinLP, emojis } = await registerMarketHelper({
+      registrant: user,
+      emojiNames: marketEmojiNames[2],
+    });
 
     const res = await Chat.submit({
       aptosConfig: aptos.config,
@@ -105,7 +104,7 @@ describe("queries swap_events and returns accurate swap row data", () => {
   });
 
   it("performs a simple liquidity fetch accurately", async () => {
-    const { marketAddress, emojicoin, emojicoinLP } = await TestHelpers.registerMarketFromNames({
+    const { marketAddress, emojicoin, emojicoinLP } = await registerMarketHelper({
       registrant: provider,
       emojiNames: marketEmojiNames[3],
     });
