+++ src/typescript/sdk/tests/e2e/queries/volume.test.ts
@@ -9,7 +9,7 @@ import { getAptosClient } from "../../utils";
 import { getFundedAccounts } from "../../utils/test-accounts";
 import { type Events } from "../../../src/emojicoin_dot_fun/events";
 import { getTxnBatchHighestVersion } from "../../utils/get-txn-batch-highest-version";
-import TestHelpers from "../../utils/helpers";
+import { registerMarketHelper } from "../../utils";
 import { fetchDailyVolumeForMarket, fetchSwapEventsBySwapper } from ".";
 
 // We need a long timeout because the test must wait for the 1-minute period to expire.
@@ -28,7 +28,7 @@ describe("queries swap_events and returns accurate swap row data", () => {
       [fundedAccounts.pop()!, 500n] as const,
     ];
     const { marketAddress, emojicoin, emojicoinLP, integrator, events } =
-      await TestHelpers.registerMarketFromNames({
+      await registerMarketHelper({
         registrant: swappersAndVolumes[0][0],
         emojiNames: ["see-no-evil monkey", "scissors"],
       });
